feat: implement URL building, header generation, and response parsing for IP data services
