+++ src/services/searchIP/AbuseIPDB.ts
@@ -1,12 +1,10 @@
 import fetch from "node-fetch";
 import type { AbuseIPObject } from "../../../types/searchIPResponse/AbuseIPDBType";
 import { ApiResponse } from "../../../types/ApiResponse/ApiResponse";
-import {
-  generateHeaders,
-  handleError,
-  parseResponse,
-} from "./simplifyFunction";
-import { isValidApiKey, buildUrl } from "./../buildUrl/buildUrl";
+import { handleError } from "./simplifyFunction";
+import { parseResponse } from "./../function/parseResponse/parseResponse";
+import { isValidApiKey, buildUrl } from "../function/buildUrl/buildUrl";
+import { generateHeaders } from "../function/generateHeaders/generateHeaders";
 
 /**
  * Fetch data from AbuseIPDB
